Move isatty closed-check snippet into stdlib_io.py
